requirements.txt
requests>=2.31.0
jinja2>=3.1.4
python-docx>=1.1.0
python-dotenv>=1.0.1
flask>=3.0.0
